+++ src/main/java/br/com/sistema/redAmber/DAO/DAOAula.java
@@ -1,6 +1,7 @@
 package br.com.sistema.redAmber.DAO;
 
 import javax.persistence.EntityManager;
+import javax.persistence.NoResultException;
 import javax.persistence.TypedQuery;
 
 import br.com.sistema.redAmber.DAO.generics.DAOGeneric;
@@ -16,14 +17,18 @@ public class DAOAula extends DAOGeneric<Aula> implements IDAOAula {
 	@Override
 	public Aula buscarAulaPorPK(AulaPK pk) {
 
-		String jpql = "SELECT aula FROM Aula aula WHERE aula.sala.id = :idSala AND aula.disciplina.id = :idDisciplina AND aula.professor.id = :idProfessor";
-		
-		TypedQuery<Aula> result = this.entityManager.createQuery(jpql, Aula.class);
-		result.setParameter("idSala", pk.getSala().getId());
-		result.setParameter("idDisciplina", pk.getDisciplina().getId());
-		result.setParameter("idProfessor", pk.getProfessor().getId());
-		
-		return result.getSingleResult();
+		try {
+			String jpql = "SELECT aula FROM Aula aula WHERE aula.id.sala.id = :idSala AND aula.id.disciplina.id = :idDisciplina AND aula.id.professor.id = :idProfessor";
+			
+			TypedQuery<Aula> result = this.entityManager.createQuery(jpql, Aula.class);
+			result.setParameter("idSala", pk.getSala().getId());
+			result.setParameter("idDisciplina", pk.getDisciplina().getId());
+			result.setParameter("idProfessor", pk.getProfessor().getId());
+			
+			return result.getSingleResult();
+		} catch (NoResultException e) {
+			return null;
+		}
 		
 	}
