+++ src/main/java/br/com/sistema/redAmber/basicas/http/HoraAulaHTTP.java
@@ -1,23 +1,12 @@
 package br.com.sistema.redAmber.basicas.http;
 
-import br.com.sistema.redAmber.basicas.enums.DiasSemana;
 import br.com.sistema.redAmber.basicas.enums.StatusHoraAula;
 
 public class HoraAulaHTTP {
 
 	private HoraAulaPKHTTP id;
 	
-	private DiasSemana dia;
 	
-	/*
-	 * TIMESTAMP
-	 */
-	private String horaInicio;
-	
-	/*
-	 * TIMESTAMP
-	 */
-	private String horaFim;
 	
 	private StatusHoraAula status;
 
@@ -31,31 +20,6 @@ public class HoraAulaHTTP {
 		this.id = id;
 	}
 
-	public DiasSemana getDia() {
-		return dia;
-	}
-
-	public void setDia(DiasSemana dia) {
-		this.dia = dia;
-	}
-
-	
-
-	public String getHoraInicio() {
-		return horaInicio;
-	}
-
-	public void setHoraInicio(String horaInicio) {
-		this.horaInicio = horaInicio;
-	}
-
-	public String getHoraFim() {
-		return horaFim;
-	}
-
-	public void setHoraFim(String horaFim) {
-		this.horaFim = horaFim;
-	}
 
 	public StatusHoraAula getStatus() {
 		return status;
