+++ src/main/java/br/com/sistema/redAmber/basicas/http/HoraAulaPKHTTP.java
@@ -1,12 +1,25 @@
 package br.com.sistema.redAmber.basicas.http;
 
 import br.com.sistema.redAmber.basicas.Turma;
+import br.com.sistema.redAmber.basicas.enums.DiasSemana;
 
 public class HoraAulaPKHTTP {
 	
 	private Turma turma;
 	
 	private AulaHTTP aula;
+	
+	private DiasSemana dia;
+	
+	/*
+	 * TIMESTAMP
+	 */
+	private String horaInicio;
+	
+	/*
+	 * TIMESTAMP
+	 */
+	private String horaFim;
 
 	public Turma getTurma() {
 		return turma;
@@ -22,6 +35,30 @@ public class HoraAulaPKHTTP {
 
 	public void setAula(AulaHTTP aula) {
 		this.aula = aula;
+	}
+
+	public DiasSemana getDia() {
+		return dia;
+	}
+
+	public void setDia(DiasSemana dia) {
+		this.dia = dia;
+	}
+
+	public String getHoraInicio() {
+		return horaInicio;
+	}
+
+	public void setHoraInicio(String horaInicio) {
+		this.horaInicio = horaInicio;
+	}
+
+	public String getHoraFim() {
+		return horaFim;
+	}
+
+	public void setHoraFim(String horaFim) {
+		this.horaFim = horaFim;
 	}
